Extract dedicated GherkinTagLine
