+++ java/src/test/java/io/cucumber/gherkin/GherkinLineTest.java
@@ -1,152 +1,18 @@
 package io.cucumber.gherkin;
 
-import io.cucumber.messages.types.Location;
 import org.junit.jupiter.api.Test;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import static java.util.Arrays.asList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class GherkinLineTest {
     private static final String WHITESPACE = "\u00A0 \t";
     private static final String CUCUMBER = "🥒";
-
-    final Location line = Locations.atLine(12);
-
-    @Test
-    void allows_any_non_space_characters_in_a_tag() {
-        GherkinLine gherkinLine = new GherkinLine("   @foo:bar  @zap\uD83E\uDD52yo", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(asList(
-                new GherkinLineSpan(4, "@foo:bar"),
-                new GherkinLineSpan(14, "@zap\uD83E\uDD52yo")
-        ), gherkinLineSpans);
-    }
-
-
-    @Test
-    void finds_tags() {
-        GherkinLine gherkinLine = new GherkinLine("@this @is @a @tag", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(asList(
-                new GherkinLineSpan(1, "@this"),
-                new GherkinLineSpan(7, "@is"),
-                new GherkinLineSpan(11, "@a"),
-                new GherkinLineSpan(14, "@tag")
-        ), gherkinLineSpans);
-    }
-
-    @Test
-    void throws_on_tags_with_spaces() {
-        GherkinLine gherkinLine = new GherkinLine("@this @is @a space separated @tag", line);
-        assertThrows(ParserException.class, gherkinLine::parseTags);
-    }
-
-    @Test
-    void throws_on_tags_with_leading_spaces() {
-        GherkinLine gherkinLine = new GherkinLine("@ leadingSpace", line);
-        assertThrows(ParserException.class, gherkinLine::parseTags);
-    }
-
-    @Test
-    void ignores_empty_tag() {
-        GherkinLine gherkinLine = new GherkinLine("@", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(Collections.emptyList(), gherkinLineSpans);
-    }
-
-    @Test
-    void ignores_empty_tags() {
-        GherkinLine gherkinLine = new GherkinLine("@@", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(Collections.emptyList(), gherkinLineSpans);
-    }
-
-    @Test
-    void finds_tags__trim_whitespace() {
-        GherkinLine gherkinLine = new GherkinLine("    @this @is  @a @tag  ", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(asList(
-                new GherkinLineSpan(5, "@this"),
-                new GherkinLineSpan(11, "@is"),
-                new GherkinLineSpan(16, "@a"),
-                new GherkinLineSpan(19, "@tag")
-        ), gherkinLineSpans);
-    }
-
-    @Test
-    void finds_tags__comment_after_tag() {
-        GherkinLine gherkinLine = new GherkinLine("@this @is #acomment  ", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(asList(
-                new GherkinLineSpan(1, "@this"),
-                new GherkinLineSpan(7, "@is")
-        ), gherkinLineSpans);
-    }
-    @Test
-    void finds_tags__comment_inside_tag_preceded_by_nbsp() {
-        GherkinLine gherkinLine = new GherkinLine("@this @is\u202F#acomment  ", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(asList(
-                new GherkinLineSpan(1, "@this"),
-                new GherkinLineSpan(7, "@is")
-        ), gherkinLineSpans);
-    }
-
-    @Test
-    void finds_tags__comment_inside_tag() {
-        GherkinLine gherkinLine = new GherkinLine("@comment_tag#2 #a comment", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(asList(
-                new GherkinLineSpan(1, "@comment_tag#2")
-        ), gherkinLineSpans);
-    }
-
-    @Test
-    void finds_tags__commented_before_tag() {
-        GherkinLine gherkinLine = new GherkinLine("@this @is #@a commented tag", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(asList(
-                new GherkinLineSpan(1, "@this"),
-                new GherkinLineSpan(7, "@is")
-        ), gherkinLineSpans);
-    }
-
-    @Test
-    void finds_tags__commented_multiple_tags() {
-        GherkinLine gherkinLine = new GherkinLine("@this @is #@a @commented @sequence of tags", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(asList(
-                new GherkinLineSpan(1, "@this"),
-                new GherkinLineSpan(7, "@is")
-        ), gherkinLineSpans);
-    }
-
-    @Test
-    void finds_tags_with_number() {
-        GherkinLine gherkinLine = new GherkinLine("@ISSUE#123", line);
-        List<GherkinLineSpan> gherkinLineSpans = gherkinLine.parseTags();
-
-        assertEquals(asList(
-                new GherkinLineSpan(1, "@ISSUE#123")
-        ), gherkinLineSpans);
-    }
 
     @Test
     void finds_table_cells() {
@@ -156,8 +22,8 @@ class GherkinLineTest {
         // - U+0009 (tab)
         // This is generated with `ruby -e 'STDOUT.write "\u00A0\u0020\u0009".encode("utf-8")' | pbcopy`
         // and pasted.
-        GherkinLine gherkinLine = new GherkinLine("      |  \tred  \t|  \tblue  \t|  \t\uD83E\uDD52\uD83E\uDD52\uD83E\uDD52  \t|  \tgreen  \t|", line);
-        
+        GherkinLine gherkinLine = new GherkinLine("      |  \tred  \t|  \tblue  \t|  \t\uD83E\uDD52\uD83E\uDD52\uD83E\uDD52  \t|  \tgreen  \t|");
+
         List<GherkinLineSpan> tableCells = gherkinLine.parseTableCells();
         GherkinLineSpan redSpan = tableCells.get(0);
         GherkinLineSpan blueSpan = tableCells.get(1);
@@ -179,7 +45,7 @@ class GherkinLineTest {
 
     @Test
     void finds_escaped_table_cells() {
-        GherkinLine gherkinLine = new GherkinLine("      | \\|æ\\\\n     | \\o\\no\\  | \\\\\\|a\\\\\\\\n | ø\\\\\\nø\\\\|", line);
+        GherkinLine gherkinLine = new GherkinLine("      | \\|æ\\\\n     | \\o\\no\\  | \\\\\\|a\\\\\\\\n | ø\\\\\\nø\\\\|");
 
         List<String> texts = gherkinLine.parseTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("|æ\\n", "\\o\no\\", "\\|a\\\\n", "ø\\\nø\\"), texts);
@@ -187,7 +53,7 @@ class GherkinLineTest {
 
     @Test
     void preserve_escaped_new_lines_at_start_and_end() {
-        GherkinLine gherkinLine = new GherkinLine("      |  \nraindrops--\nher last kiss\ngoodbye.\n  |", line);
+        GherkinLine gherkinLine = new GherkinLine("      |  \nraindrops--\nher last kiss\ngoodbye.\n  |");
         List<String> texts = gherkinLine.parseTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("" +
                 "\n" +
@@ -199,42 +65,42 @@ class GherkinLineTest {
 
     @Test
     void escapes_backslash() {
-        GherkinLine gherkinLine = new GherkinLine("|\\\\o\\no\\||", line);
+        GherkinLine gherkinLine = new GherkinLine("|\\\\o\\no\\||");
         List<String> texts = gherkinLine.parseTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("\\o\no|"), texts);
     }
 
     @Test
     void throws_on_illegal_escapes_backslash() {
-        GherkinLine gherkinLine = new GherkinLine("|\\o\\no\\||", line);
+        GherkinLine gherkinLine = new GherkinLine("|\\o\\no\\||");
         List<String> texts = gherkinLine.parseTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("\\o\no|"), texts);
     }
 
     @Test
     void correctly_trims_white_spaces_before_cell_content() {
-        GherkinLine gherkinLine = new GherkinLine("|   \t spaces before|", line);
+        GherkinLine gherkinLine = new GherkinLine("|   \t spaces before|");
         List<String> texts = gherkinLine.parseTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("spaces before"), texts);
     }
 
     @Test
     void correctly_trims_white_spaces_after_cell_content() {
-        GherkinLine gherkinLine = new GherkinLine("|spaces after   |", line);
+        GherkinLine gherkinLine = new GherkinLine("|spaces after   |");
         List<String> texts = gherkinLine.parseTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("spaces after"), texts);
     }
 
     @Test
     void correctly_trims_white_spaces_around_cell_content() {
-        GherkinLine gherkinLine = new GherkinLine("|   \t spaces everywhere   \t|", line);
+        GherkinLine gherkinLine = new GherkinLine("|   \t spaces everywhere   \t|");
         List<String> texts = gherkinLine.parseTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("spaces everywhere"), texts);
     }
 
     @Test
     void does_not_drop_white_spaces_inside_a_cell() {
-        GherkinLine gherkinLine = new GherkinLine("| foo()\n  bar\nbaz |", line);
+        GherkinLine gherkinLine = new GherkinLine("| foo()\n  bar\nbaz |");
         List<String> texts = gherkinLine.parseTableCells().stream().map(span -> span.text).collect(Collectors.toList());
         assertEquals(asList("foo()\n  bar\nbaz"), texts);
     }
@@ -242,7 +108,7 @@ class GherkinLineTest {
     @Test
     void startsWithTitleKeyword_corresponding_keyword_match() {
         // Given
-        GherkinLine gherkinLine = new GherkinLine("Feature: Hello", line);
+        GherkinLine gherkinLine = new GherkinLine("Feature: Hello");
 
         // When/Then
         assertTrue(gherkinLine.startsWithTitleKeyword("Feature"));
@@ -251,7 +117,7 @@ class GherkinLineTest {
     @Test
     void startsWithTitleKeyword_non_corresponding_keyword_does_not_match() {
         // Given
-        GherkinLine gherkinLine = new GherkinLine("Rule: X", line);
+        GherkinLine gherkinLine = new GherkinLine("Rule: X");
 
         // When/Then
         assertFalse(gherkinLine.startsWithTitleKeyword("Background")); // not the same keyword
@@ -262,25 +128,16 @@ class GherkinLineTest {
     @Test
     void startsWithTitleKeyword_keyword_without_colon_does_not_match() {
         // Given
-        GherkinLine gherkinLine = new GherkinLine("Rule", line);
+        GherkinLine gherkinLine = new GherkinLine("Rule");
 
         // When/Then
         assertFalse(gherkinLine.startsWithTitleKeyword("Rule"));
     }
 
     @Test
-    void parseTags_returns_empty_list_when_empty_line() {
-        // Given
-        GherkinLine gherkinLine = new GherkinLine("", line);
-
-        // When/Then
-        assertTrue(gherkinLine.parseTags().isEmpty());
-    }
-
-    @Test
     void substringTrimmed_empty_idempotence() {
         // Given
-        GherkinLine gherkinLine = new GherkinLine("", line);
+        GherkinLine gherkinLine = new GherkinLine("");
 
         // When/Then
         assertEquals("", gherkinLine.substringTrimmed(0));
@@ -289,7 +146,7 @@ class GherkinLineTest {
     @Test
     void substringTrimmed_regular_spaces_after_index_are_trimmed() {
         // Given
-        GherkinLine gherkinLine = new GherkinLine("Rule:    my rule name ", line);
+        GherkinLine gherkinLine = new GherkinLine("Rule:    my rule name ");
 
         // When/Then
         assertEquals("my rule name", gherkinLine.substringTrimmed(5));
@@ -298,7 +155,7 @@ class GherkinLineTest {
     @Test
     void substringTrimmed_unicode_spaces_after_index_are_trimmed() {
         // Given
-        GherkinLine gherkinLine = new GherkinLine("Rule:" + WHITESPACE + CUCUMBER + WHITESPACE, line);
+        GherkinLine gherkinLine = new GherkinLine("Rule:" + WHITESPACE + CUCUMBER + WHITESPACE);
 
         // When/Then
         assertEquals(CUCUMBER, gherkinLine.substringTrimmed(5));
@@ -307,7 +164,7 @@ class GherkinLineTest {
     @Test
     void getRawTextSubstring_substring_from_beginning() {
         // Given
-        GherkinLine gherkinLine = new GherkinLine("    some text", line);
+        GherkinLine gherkinLine = new GherkinLine("    some text");
 
         // When/Then
         assertEquals("    some text", gherkinLine.getRawTextSubstring(0));
@@ -316,7 +173,7 @@ class GherkinLineTest {
     @Test
     void getRawTextSubstring_substring_from_middle() {
         // Given
-        GherkinLine gherkinLine = new GherkinLine("    some text", line);
+        GherkinLine gherkinLine = new GherkinLine("    some text");
 
         // When/Then
         assertEquals(" some text", gherkinLine.getRawTextSubstring(3));
